remove header and just define structs inside
